AIandPrompt/kpl_ai_backend_deepseek/requirements.txt
openai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
cachetools>=5.3.0